Time: 0 ms (100%), Space: 13.7 MB (47.87%) - LeetHub

+++ 1950-sign-of-the-product-of-an-array/1950-sign-of-the-product-of-an-array.cpp
@@ -7,7 +7,7 @@ public:
         return 0;
     }
     int arraySign(vector<int>& nums) {
-        long long p=1;
+        int p=1;
         for(int i=0 ; i<nums.size() ; i++){
             if(nums[i] < 0)p*=-1;
             if(nums[i] == 0)p=0;
